pyproject.toml
[project]
name = "cloning-eval"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.57.1",
    "beautifulsoup4>=4.13.4",
    "htmlmin>=0.1.12",
    "ipykernel>=6.29.5",
    "pandas>=2.3.1",
    "pillow>=11.3.0",
    "playwright>=1.53.0",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "selenium>=4.34.2",
    "tenacity>=9.1.2",
]
